refactor(components.uirouter.line.progress): remove ng-ovh prefix
Since this component could used on a multiple projects,
it doesn't make sense to add the `ng-ovh` prefix.

As same as ng-uirouter-title component (which could be moved later to this monorepo).

ref: https://github.com/ovh-ux/ng-uirouter-title.
