--- packages/components/ng-ovh-uirouter-line-progress/README.md
@@ -1,35 +1,0 @@
-# ng-ovh-uirouter-line-progress
-
-
-[![Downloads](https://badgen.net/npm/dt/@ovh-ux/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress) [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/manager/modules/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress?activeTab=dependencies) [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/manager/modules/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress?activeTab=dependencies) [![Gitter](https://badgen.net/badge/gitter/ovh-ux/blue?icon=gitter)](https://gitter.im/ovh/ux)
-
-## Install
-
-```sh
-yarn add @ovh-ux/ng-ovh-uirouter-line-progress
-```
-## Usage
-
-```js
-import angular from 'angular';
-import ngUiRouterLineProgress from '@ovh-ux/ng-ovh-uirouter-line-progress';
-
-angular
-  .module('myApp', [
-    ngUiRouterLineProgress,
-  ]);
-```
-
-## Test
-
-```sh
-yarn test
-```
-
-## Contributing
-
-Always feel free to help out! Whether it's [filing bugs and feature requests](https://github.com/ovh-ux/manager/issues/new) or working on some of the [open issues](https://github.com/ovh-ux/manager/issues), our [contributing guide](CONTRIBUTING.md) will help get you started.
-
-## License
-
-[BSD-3-Clause](LICENSE) © OVH SAS
