+++ packages/components/ng-uirouter-line-progress/package.json
@@ -1,15 +1,18 @@
 {
-  "name": "@ovh-ux/ng-ovh-uirouter-line-progress",
+  "name": "@ovh-ux/ng-uirouter-line-progress",
   "version": "0.0.0",
   "description": "UiRouter line progress",
   "keywords": [
-    "uirouter",
-    "ovh"
+    "angular",
+    "nprogress",
+    "ovh",
+    "progress",
+    "uirouter"
   ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ovh-ux/manager.git",
-    "directory": "packages/components/ng-ovh-uirouter-line-progress"
+    "directory": "packages/components/ng-uirouter-line-progress"
   },
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
@@ -18,15 +21,15 @@
   ],
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
-  "browser": "./dist/umd/ng-ovh-uirouter-line-progress.js",
+  "browser": "./dist/umd/ng-uirouter-line-progress.js",
   "scripts": {
     "build": "rollup -c --environment BUILD:production",
     "dev": "rollup -c --environment BUILD:development",
     "dev:watch": "yarn run dev --watch",
     "prepare": "yarn run build",
-    "start": "lerna exec --stream --scope='@ovh-ux/ng-ovh-uirouter-line-progress' --include-filtered-dependencies -- yarn run build",
-    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-ovh-uirouter-line-progress' --include-filtered-dependencies -- yarn run dev",
-    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-ovh-uirouter-line-progress' --include-filtered-dependencies -- yarn run dev:watch"
+    "start": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-line-progress' --include-filtered-dependencies -- yarn run build",
+    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-line-progress' --include-filtered-dependencies -- yarn run dev",
+    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-uirouter-line-progress' --include-filtered-dependencies -- yarn run dev:watch"
   },
   "dependencies": {
     "nprogress": "^0.2.0"
