+++ packages/components/ng-uirouter-line-progress/rollup.config.js
