+++ packages/components/ng-uirouter-line-progress/src/index.js
@@ -1,9 +1,9 @@
 import angular from 'angular';
 import '@uirouter/angularjs';
 
-import config from './ng-ovh-uirouter-line-progress.config';
+import config from './ng-uirouter-line-progress.config';
 
-import './ng-ovh-uirouter-line-progress.less';
+import './ng-uirouter-line-progress.less';
 
 const moduleName = 'ngUirouterLineProgress';
