+++ packages/components/ng-uirouter-line-progress/src/ng-uirouter-line-progress.config.js
