+++ packages/manager/apps/public-cloud/package.json
@@ -36,10 +36,10 @@
     "@ovh-ux/ng-ovh-swimming-poll": "^4.0.0",
     "@ovh-ux/ng-ovh-toaster": "^1.0.2",
     "@ovh-ux/ng-ovh-uirouter-layout": "^0.0.0",
-    "@ovh-ux/ng-ovh-uirouter-line-progress": "0.0.0",
     "@ovh-ux/ng-ovh-user-pref": "^1.0.0",
     "@ovh-ux/ng-translate-async-loader": "^2.0.0",
     "@ovh-ux/ng-uirouter-breadcrumb": "0.0.0",
+    "@ovh-ux/ng-uirouter-line-progress": "0.0.0",
     "@uirouter/angularjs": "^1.0.20",
     "angular": "^1.7.5",
     "angular-animate": "^1.7.5",
