+++ packages/manager/apps/public-cloud/src/index.js
@@ -23,7 +23,7 @@ import ovhManagerPci from '@ovh-ux/manager-pci';
 import ngOvhApiWrappers from '@ovh-ux/ng-ovh-api-wrappers';
 import ngOvhUserPref from '@ovh-ux/ng-ovh-user-pref';
 import ngUiRouterBreadcrumb from '@ovh-ux/ng-uirouter-breadcrumb';
-import ngUiRouterLineProgress from '@ovh-ux/ng-ovh-uirouter-line-progress';
+import ngUiRouterLineProgress from '@ovh-ux/ng-uirouter-line-progress';
 
 import 'ovh-ui-kit/dist/oui.css';
 import 'ovh-ui-kit-bs/dist/ovh-ui-kit-bs.css';
